refactor(point): pointService 구현체 분리
